+++ nautobot_ssot_device42/diffsync/from_d42/models/dcim.py
@@ -26,7 +26,7 @@ from nautobot.extras.choices import CustomFieldTypeChoices
 from nautobot.extras.models import CustomField, Relationship, RelationshipAssociation
 from nautobot.extras.models import Status as NautobotStatus
 from nautobot.ipam.models import VLAN as NautobotVLAN
-from nautobot_ssot_device42.constant import DEFAULTS, INTF_SPEED_MAP
+from nautobot_ssot_device42.constant import DEFAULTS, INTF_SPEED_MAP, PLUGIN_CFG
 from nautobot_ssot_device42.utils import device42, nautobot
 
 try:
@@ -128,11 +128,12 @@ class Building(DiffSyncModel):
         The self.diffsync.objects_to_delete dictionary stores all objects for deletion and removes them from Nautobot
         in the correct order. This is used in the Nautobot adapter sync_complete function.
         """
-        super().delete()
-        site = NautobotSite.objects.get(id=self.uuid)
-        if self.diffsync.job.debug:
-            self.diffsync.job.log_warning(message=f"Site {self.name} will be deleted.")
-        self.diffsync.objects_to_delete["site"].append(site)  # pylint: disable=protected-access
+        if PLUGIN_CFG.get("delete_on_sync"):
+            super().delete()
+            site = NautobotSite.objects.get(id=self.uuid)
+            if self.diffsync.job.debug:
+                self.diffsync.job.log_warning(message=f"Site {self.name} will be deleted.")
+            self.diffsync.objects_to_delete["site"].append(site)  # pylint: disable=protected-access
         return self
 
 
@@ -193,11 +194,12 @@ class Room(DiffSyncModel):
 
     def delete(self):
         """Delete RackGroup object from Nautobot."""
-        super().delete()
-        rackgroup = NautobotRackGroup.objects.get(self.uuid)
-        if self.diffsync.job.debug:
-            self.diffsync.job.log_warning(message=f"RackGroup {self.name} will be deleted.")
-        rackgroup.delete()
+        if PLUGIN_CFG.get("delete_on_sync"):
+            super().delete()
+            rackgroup = NautobotRackGroup.objects.get(self.uuid)
+            if self.diffsync.job.debug:
+                self.diffsync.job.log_warning(message=f"RackGroup {self.name} will be deleted.")
+            rackgroup.delete()
         return self
 
 
@@ -277,11 +279,12 @@ class Rack(DiffSyncModel):
         The self.diffsync.objects_to_delete dictionary stores all objects for deletion and removes them from Nautobot
         in the correct order. This is used in the Nautobot adapter sync_complete function.
         """
-        super().delete()
-        rack = NautobotRack.objects.get(id=self.uuid)
-        if self.diffsync.job.debug:
-            self.diffsync.job.log_warning(message=f"Rack {self.name} will be deleted.")
-        self.diffsync.objects_to_delete["rack"].append(rack)  # pylint: disable=protected-access
+        if PLUGIN_CFG.get("delete_on_sync"):
+            super().delete()
+            rack = NautobotRack.objects.get(id=self.uuid)
+            if self.diffsync.job.debug:
+                self.diffsync.job.log_warning(message=f"Rack {self.name} will be deleted.")
+            self.diffsync.objects_to_delete["rack"].append(rack)  # pylint: disable=protected-access
         return self
 
 
@@ -344,11 +347,12 @@ class Vendor(DiffSyncModel):
         The self.diffsync.objects_to_delete dictionary stores all objects for deletion and removes them from Nautobot
         in the correct order. This is used in the Nautobot adapter sync_complete function.
         """
-        super().delete()
-        _manu = NautobotManufacturer.objects.get(id=self.uuid)
-        if self.diffsync.job.debug:
-            self.diffsync.job.log_warning(message=f"Manufacturer {self.name} will be deleted.")
-        self.diffsync.objects_to_delete["manufacturer"].append(_manu)  # pylint: disable=protected-access
+        if PLUGIN_CFG.get("delete_on_sync"):
+            super().delete()
+            _manu = NautobotManufacturer.objects.get(id=self.uuid)
+            if self.diffsync.job.debug:
+                self.diffsync.job.log_warning(message=f"Manufacturer {self.name} will be deleted.")
+            self.diffsync.objects_to_delete["manufacturer"].append(_manu)  # pylint: disable=protected-access
         return self
 
 
@@ -427,11 +431,12 @@ class Hardware(DiffSyncModel):
         The self.diffsync.objects_to_delete dictionary stores all objects for deletion and removes them from Nautobot
         in the correct order. This is used in the Nautobot adapter sync_complete function.
         """
-        super().delete()
-        _dt = NautobotDeviceType.objects.get(id=self.uuid)
-        if self.diffsync.job.debug:
-            self.diffsync.job.log_warning(message=f"DeviceType {self.name} will be deleted.")
-        self.diffsync.objects_to_delete["device_type"].append(_dt)  # pylint: disable=protected-access
+        if PLUGIN_CFG.get("delete_on_sync"):
+            super().delete()
+            _dt = NautobotDeviceType.objects.get(id=self.uuid)
+            if self.diffsync.job.debug:
+                self.diffsync.job.log_warning(message=f"DeviceType {self.name} will be deleted.")
+            self.diffsync.objects_to_delete["device_type"].append(_dt)  # pylint: disable=protected-access
         return self
 
 
@@ -523,11 +528,12 @@ class Cluster(DiffSyncModel):
         The self.diffsync.objects_to_delete dictionary stores all objects for deletion and removes them from Nautobot
         in the correct order. This is used in the Nautobot adapter sync_complete function.
         """
-        super().delete()
-        _cluster = NautobotVC.objects.get(id=self.uuid)
-        if self.diffsync.job.debug:
-            self.diffsync.job.log_warning(message=f"Virtual Chassis {self.name} will be deleted.")
-        self.diffsync.objects_to_delete["cluster"].append(_cluster)  # pylint: disable=protected-access
+        if PLUGIN_CFG.get("delete_on_sync"):
+            super().delete()
+            _cluster = NautobotVC.objects.get(id=self.uuid)
+            if self.diffsync.job.debug:
+                self.diffsync.job.log_warning(message=f"Virtual Chassis {self.name} will be deleted.")
+            self.diffsync.objects_to_delete["cluster"].append(_cluster)  # pylint: disable=protected-access
         return self
 
 
@@ -805,11 +811,12 @@ class Device(DiffSyncModel):
         The self.diffsync.objects_to_delete dictionary stores all objects for deletion and removes them from Nautobot
         in the correct order. This is used in the Nautobot adapter sync_complete function.
         """
-        super().delete()
-        _dev = NautobotDevice.objects.get(id=self.uuid)
-        if self.diffsync.job.debug:
-            self.diffsync.job.log_warning(message=f"Device {self.name} will be deleted.")
-        self.diffsync.objects_to_delete["device"].append(_dev)  # pylint: disable=protected-access
+        if PLUGIN_CFG.get("delete_on_sync"):
+            super().delete()
+            _dev = NautobotDevice.objects.get(id=self.uuid)
+            if self.diffsync.job.debug:
+                self.diffsync.job.log_warning(message=f"Device {self.name} will be deleted.")
+            self.diffsync.objects_to_delete["device"].append(_dev)  # pylint: disable=protected-access
         return self
 
     @staticmethod
@@ -990,11 +997,12 @@ class Port(DiffSyncModel):
 
     def delete(self):
         """Delete Interface object from Nautobot."""
-        super().delete()
-        _intf = NautobotInterface.objects.get(id=self.uuid)
-        if self.diffsync.job.debug:
-            self.diffsync.job.log_warning(message=f"Interface {self.name} for {self.device} will be deleted.")
-        self.diffsync.objects_to_delete["port"].append(_intf)  # pylint: disable=protected-access
+        if PLUGIN_CFG.get("delete_on_sync"):
+            super().delete()
+            _intf = NautobotInterface.objects.get(id=self.uuid)
+            if self.diffsync.job.debug:
+                self.diffsync.job.log_warning(message=f"Interface {self.name} for {self.device} will be deleted.")
+            self.diffsync.objects_to_delete["port"].append(_intf)  # pylint: disable=protected-access
         return self
 
 
@@ -1181,11 +1189,12 @@ class Connection(DiffSyncModel):
 
     def delete(self):
         """Delete Cable object from Nautobot."""
-        super().delete()
-        _conn = NautobotCable.objects.get(id=self.uuid)
-        if self.diffsync.job.debug:
-            self.diffsync.job.log_info(message=f"Deleting Cable {self.get_identifiers()} UUID: {self.uuid}")
-        _conn.delete()
+        if PLUGIN_CFG.get("delete_on_sync"):
+            super().delete()
+            _conn = NautobotCable.objects.get(id=self.uuid)
+            if self.diffsync.job.debug:
+                self.diffsync.job.log_info(message=f"Deleting Cable {self.get_identifiers()} UUID: {self.uuid}")
+            _conn.delete()
         return self
